+++ apps/fluent-tester/src/TestComponents/StrokeWidth/consts.ts
@@ -1,0 +1,2 @@
+export const HOMEPAGE_STROKEWIDTH_BUTTON = 'Homepage_StrokeWidth_Button';
+export const STROKEWIDTH_TESTPAGE = 'StrokeWidth_TestPage';