+++ packages/experimental/ExperimentalIcon/src/Icon.types.ts
@@ -1,0 +1,7 @@
+import { SvgIconProps } from './SvgIcon/SvgIcon.types';
+import { FontIconProps } from './FontIcon/FontIcon.types';
+
+export interface IconProps {
+  svgSource?: SvgIconProps;
+  fontSource?: FontIconProps;
+}